# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Telegram AI Autonomous System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all Telegram accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TelegramAccount"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Add a new Telegram account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramAccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramAccount"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramAccount"
    patch:
      operationId: updateAccount
      tags: [accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramAccountUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramAccount"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /accounts/{id}/connect:
    post:
      operationId: connectAccount
      tags: [accounts]
      summary: Connect / authenticate a Telegram account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountConnectInput"
      responses:
        "200":
          description: Connection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountConnectResult"
  /accounts/{id}/disconnect:
    post:
      operationId: disconnectAccount
      tags: [accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountConnectResult"
...
  /ai/generate:
    post:
      operationId: generateContent
      tags: [ai]
      summary: Generate AI content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIGenerateInput"
      responses:
        "200":
          description: Generated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIGenerateResult"
  /ai/personalities:
    get:
      operationId: listPersonalities
      tags: [ai]
      summary: List AI personalities
      responses:
        "200":
          description: Personalities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AIPersonality"
    post:
      operationId: createPersonality
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIPersonalityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIPersonality"
  /ai/personalities/{id}:
    patch:
      operationId: updatePersonality
      tags: [ai]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIPersonalityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIPersonality"
    delete:
      operationId: deletePersonality
      tags: [ai]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    TelegramAccount:
      type: object
      required: [id, phoneNumber, username, status, personalityId, isConnected, createdAt]
      properties:
        id:
          type: integer
        phoneNumber:
          type: string
        username:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, banned, pending, cooldown]
        personalityId:
          type: ["integer", "null"]
        isConnected:
          type: boolean
        lastSeen:
          type: ["string", "null"]
        messagesCount:
          type: integer
        groupsCount:
          type: integer
        healthScore:
          type: integer
        timezone:
          type: ["string", "null"]
        activeHoursStart:
          type: ["integer", "null"]
        activeHoursEnd:
          type: ["integer", "null"]
        createdAt:
          type: string
    TelegramAccountInput:
      type: object
      required: [phoneNumber]
      properties:
        phoneNumber:
          type: string
        displayName:
          type: string
        personalityId:
          type: integer
        timezone:
          type: string
        activeHoursStart:
          type: integer
        activeHoursEnd:
          type: integer
    TelegramAccountUpdate:
      type: object
      properties:
        displayName:
          type: string
        personalityId:
          type: integer
        status:
          type: string
        timezone:
          type: string
        activeHoursStart:
          type: integer
        activeHoursEnd:
          type: integer
    AccountConnectInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [request_code, verify_code, verify_2fa]
        code:
          type: string
        password:
          type: string
    AccountConnectResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        needsCode:
          type: boolean
        needs2FA:
          type: boolean
...
    AIPersonality:
      type: object
      required: [id, name, type, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [professional, business, crypto, motivational, friendly, community, support, casual]
        description:
          type: ["string", "null"]
        tone:
          type: string
        emojiFrequency:
          type: string
          enum: [none, low, medium, high]
        energyLevel:
          type: integer
        isDefault:
          type: boolean
        createdAt:
          type: string
    AIPersonalityInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [professional, business, crypto, motivational, friendly, community, support, casual]
        description:
          type: string
        tone:
          type: string
        emojiFrequency:
          type: string
          enum: [none, low, medium, high]
        energyLevel:
          type: integer
    AIPersonalityUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        tone:
          type: string
        emojiFrequency:
          type: string
        energyLevel:
          type: integer
        isDefault:
          type: boolean
    AIGenerateInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
        personalityId:
          type: integer
        contentType:
          type: string
          enum: [motivation, announcement, marketing, welcome, reply, custom]
        language:
          type: string
        context:
          type: string
    AIGenerateResult:
      type: object
      required: [content, model]
      properties:
        content:
          type: string
        model:
          type: string
        tokensUsed:
          type: integer
        personalityApplied:
          type: ["string", "null"]
...